+++ src/panels/panels.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react';
 
-import { Typography, Box, CircularProgress, Tabs, Tab, Icon, createMuiTheme, ThemeProvider } from '@material-ui/core';
-import { withTranslation } from 'react-i18next';
+import { Typography, Box, CircularProgress, Tabs, Tab, Icon, createMuiTheme, ThemeProvider, makeStyles, createStyles, useMediaQuery } from '@material-ui/core';
+import { useTranslation } from 'react-i18next';
 
 const PanelSave = React.lazy(() => import(/* webpackChunkName: "panelSave" */ './save'));
 const PanelLayout = React.lazy(() => import(/* webpackChunkName: "panelLayout" */ './panel-layout'));
@@ -9,27 +9,24 @@ const PanelDesign = React.lazy(() => import(/* webpackChunkName: "panelDesign" *
 const PanelStations = React.lazy(() => import(/* webpackChunkName: "panelStations" */ './stations'));
 const PanelInfo = React.lazy(() => import(/* webpackChunkName: "panelInfo" */ './panel-info'));
 
+const panelStyles = makeStyles(theme => (
+    createStyles({
+        root: {
+            background: theme.palette.background.default,
+        }
+    })
+));
+
 interface PanelProps {
     children?: React.ReactNode;
-    index: any;
     value: any;
 }
 
-interface PanelState {
-    panel: JSX.Element;
-}
+function Panel(props: PanelProps) {
+    const classes = panelStyles();
 
-class Panel extends React.Component<PanelProps, PanelState> {
-    constructor(props) {
-        super(props);
-
-        this.state = {
-            panel: this.panel
-        }
-    }
-
-    get panel() {
-        switch (this.props.value) {
+    const panel = (index: number) => {
+        switch (index) {
             case 0:
                 return <PanelSave />;
             case 1:
@@ -43,46 +40,41 @@ class Panel extends React.Component<PanelProps, PanelState> {
             default:
                 return <PanelSave />;
         };
-    }
+    };
 
-    componentDidUpdate(prevProps) {
-        if (prevProps.value === this.props.value) {return;}
-        this.setState({panel: this.panel});
-    }
+    const { children, value, ...other } = props;
+    return (
+        <Typography
+            className={classes.root}
+            component="div"
+            role="tabpanel"
+            {...other}>
+            <Box p={3} style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
+                <React.Suspense fallback={<CircularProgress />}>
+                    {panel(props.value)}
+                </React.Suspense>
+            </Box>
+        </Typography>
+    );
 
-    render() {
-        // const { children, value, index, ...other } = this.props;
-        // return (
-        //     <Typography
-        //       component="div"
-        //       role="tabpanel"
-        //       hidden={value !== index}
-        //       id={`simple-tabpanel-${index}`}
-        //       aria-labelledby={`simple-tab-${index}`}
-        //       {...other}
-        //     >
-        //       {value === index && <Box p={3}>{children}</Box>}
-        //     </Typography>
-        // );
-        const { children, value, index, ...other } = this.props;
-        return (
-            <Typography
-                component="div"
-                role="tabpanel"
-                id={`simple-tabpanel-${index}`}
-                aria-labelledby={`simple-tab-${index}`}
-                {...other}>
-                <Box p={3} style={{display: 'flex', justifyContent: 'center', alignItems: 'center'}}>
-                    <React.Suspense fallback={<CircularProgress />}>
-                        {this.state.panel}
-                    </React.Suspense>
-                </Box>
-            </Typography>
-        )
-    }
 }
 
-const theme = createMuiTheme({
+const darkTheme = createMuiTheme({
+    palette: {
+        type: 'dark',
+        primary: {
+            main: '#bb86fc',
+        },
+        secondary: {
+            main: '#018786',
+        },
+        error: {
+            main: '#cf6679',
+        }, 
+    },
+});
+
+const lightTheme = createMuiTheme({
     palette: {
         primary: {
             light: '#bb86fc',
@@ -93,75 +85,59 @@ const theme = createMuiTheme({
         },
         error: {
             main: '#b00020',
-            dark: '#cf6679',
         }, 
     },
 });
 
-interface PanelTabProps {
-    t: any;
-}
+export default function PanelTab(props) {
+    const { t } = useTranslation('', { useSuspense: false });
 
-interface PanelTabState {
-    value: number;
-}
+    const prefersDarkMode = useMediaQuery('(prefers-color-scheme: dark)');
+    const theme = prefersDarkMode ? darkTheme : lightTheme;
 
-class PanelTab extends React.Component<PanelTabProps, PanelTabState> {
-    constructor(props) {
-        super(props);
-        this.state = {
-            value: 0
-        };
-        this.handleChange = this.handleChange.bind(this);
-    }
+    const [value, setValue] = React.useState(0);
 
-    handleChange(event: React.ChangeEvent<{}>, newValue: number) {
-        this.setState({value: newValue})
+    const tabClasses = {
+        root: 'tab-nav-main',
+        wrapper: 'tab-nav', 
+        labelIcon: 'tab-nav',
     };
-
-    render() {
-        let tabClasses = {
-            root: 'tab-nav-main',
-            wrapper: 'tab-nav', 
-            labelIcon: 'tab-nav',
-        };
-        return (
-            <div>
-                <ThemeProvider theme={theme}>
-                    <div>
-                        <Tabs value={this.state.value} indicatorColor="primary" textColor="primary" onChange={this.handleChange} variant="scrollable" scrollButtons="off">
-                            <Tab 
-                                label={<span>{this.props.t('tab.file')}</span>}
-                                icon={<Icon>insert_drive_file</Icon>}
-                                classes={tabClasses}
-                            />
-                            <Tab 
-                                label={<span>{this.props.t('tab.layout')}</span>}
-                                icon={<Icon>panorama</Icon>}
-                                classes={tabClasses}
-                            />
-                            <Tab 
-                                label={<span>{this.props.t('tab.design')}</span>}
-                                icon={<Icon>brush</Icon>}
-                                classes={tabClasses}
-                            />
-                            <Tab
-                                label={<span>{this.props.t('tab.stations')}</span>}
-                                icon={<Icon>directions_transit</Icon>}
-                                classes={tabClasses}
-                            />
-                            <Tab
-                                label={<span>{this.props.t('tab.info')}</span>}
-                                icon={<Icon>info</Icon>}
-                                classes={tabClasses}
-                            />
-                        </Tabs>
-                    </div>
-                    <Panel value={this.state.value} index={0} />
-                </ThemeProvider>
-            </div>
-        );
-    }
+    return (
+        <div>
+            <ThemeProvider theme={theme}>
+                <Typography style={{background: theme.palette.background.default}}>
+                    <Tabs value={value} indicatorColor="primary" 
+                        textColor="primary" onChange={(e, val) => setValue(val)} 
+                        variant="scrollable" scrollButtons="off">
+                        <Tab 
+                            label={<span>{t('tab.file')}</span>}
+                            icon={<Icon>insert_drive_file</Icon>}
+                            classes={tabClasses}
+                        />
+                        <Tab 
+                            label={<span>{t('tab.layout')}</span>}
+                            icon={<Icon>panorama</Icon>}
+                            classes={tabClasses}
+                        />
+                        <Tab 
+                            label={<span>{t('tab.design')}</span>}
+                            icon={<Icon>brush</Icon>}
+                            classes={tabClasses}
+                        />
+                        <Tab
+                            label={<span>{t('tab.stations')}</span>}
+                            icon={<Icon>directions_transit</Icon>}
+                            classes={tabClasses}
+                        />
+                        <Tab
+                            label={<span>{t('tab.info')}</span>}
+                            icon={<Icon>info</Icon>}
+                            classes={tabClasses}
+                        />
+                    </Tabs>
+                </Typography>
+                <Panel value={value} />
+            </ThemeProvider>
+        </div>
+    );
 }
-
-export default withTranslation()(PanelTab);