+++ src/panels/stations/chip-set.tsx
@@ -1,10 +1,28 @@
 import * as React from 'react';
 import { useTranslation } from 'react-i18next';
-import { Chip, makeStyles, Avatar, Icon, createStyles } from '@material-ui/core';
+import { Chip, makeStyles, Avatar, Icon, createStyles, useMediaQuery } from '@material-ui/core';
 import { StationInfo } from '../../types';
 
 const useStyles = makeStyles(theme => {
     return createStyles({
+        root: {
+            backgroundColor: theme.palette.background.paper, 
+            height: 55, 
+            borderRadius: 8, 
+            boxShadow: theme.shadows[1]
+        },
+        rootSize: {
+            margin: 8, 
+            width: 'calc((100% - 96px) / 6)', 
+        }, 
+        rootSizeTablet: {
+            margin: 6, 
+            width: 'calc((100% - 48px) / 4)', 
+        }, 
+        rootSizeMobile: {
+            margin: 6, 
+            width: 'calc((100% - 24px) / 2)', 
+        }, 
         stnChipText: {
             display: 'block',
             textAlign: 'center' as 'center'
@@ -21,6 +39,16 @@ const useStyles = makeStyles(theme => {
             lineHeight: '.9rem', 
             whiteSpace: 'pre' as 'pre', 
             overflow: 'hidden'
+        }, 
+        addChipLabel: {
+            fontWeight: 500, 
+            fontSize: '1rem', 
+            lineHeight: '1.2rem', 
+            textAlign: 'center', 
+            width: '100%', 
+        }, 
+        stnChipLabel: {
+            width: '100%'
         }
     });
 })
@@ -35,6 +63,8 @@ interface StationChipSetProps {
 
 export default function StationChipSet(props: StationChipSetProps) {
     const {t, i18n} = useTranslation();
+    const isTablet = useMediaQuery('(max-width: 839px) and (min-width: 480px)');
+    const isMobile = useMediaQuery('(max-width: 480px)');
     const classes = useStyles();
     return (
         <div>
@@ -42,8 +72,12 @@ export default function StationChipSet(props: StationChipSetProps) {
                 key="add_stn"
                 icon={<Avatar style={{backgroundColor: 'unset'}}><Icon>add_box</Icon></Avatar>}
                 label={t('stations.add.button')}
-                className={'stn-chip'}
-                classes={{label: 'stn-chip__label'}}
+                className={
+                    `${classes.root} ${
+                        isMobile ? classes.rootSizeMobile : (isTablet ? classes.rootSizeTablet : classes.rootSize)
+                    }`
+                }
+                classes={{label: classes.addChipLabel}}
                 onClick={props.addStationClick}
             />
             {window.myLine.tpo.map(stnId => {
@@ -64,7 +98,12 @@ export default function StationChipSet(props: StationChipSetProps) {
                             <Avatar style={{backgroundColor: 'unset'}}>{props.stnList[stnId].num}</Avatar>}
                         label={label}
                         onClick={() => props.onSelection(stnId)}
-                        className={'stn-chip'} />
+                        classes={{label: classes.stnChipLabel}}
+                        className={
+                            `${classes.root} ${
+                                isMobile ? classes.rootSizeMobile : (isTablet ? classes.rootSizeTablet : classes.rootSize)
+                            }`
+                        } />
                 )
             })}
         </div>